requirements.txt
beautifulsoup4==4.10.0
Pillow==8.4.0
PyQt5==5.15.6
requests==2.25.1
selenium==3.141.0
